✨feat: 최종 라운드 결과 & 최종 결과 로직 분리
