+++ src/components/BalanceGameProgress/BalanceGameProgress.tsx
@@ -8,21 +8,18 @@ import * as StompJS from '@stomp/stompjs';
 import { SOCKET } from '@/constants/websocket';
 import { getBalanceGameResults } from '@/services/balanceGames';
 import { BalanceGameResultGetResponse } from '@/types/api';
-import { BarProps } from '@/components/FinalResultChart/Bar';
 
 interface BalanceGameProgressProps {
   sendMessage: <T>(
     params: Omit<StompJS.IPublishParams, 'body'> & { body?: T }
   ) => void;
   setResult: (result: BalanceGameResultGetResponse[]) => void;
-  setFinalResult: (result: BarProps[]) => void;
   roomId: string;
 }
 
 const BalanceGameProgress = ({
   sendMessage,
   setResult,
-  setFinalResult,
   roomId,
 }: BalanceGameProgressProps) => {
   const { round, setRoomStatus } = useBalanceGameStore();
@@ -35,34 +32,18 @@ const BalanceGameProgress = ({
       setRoomStatus('result');
       getBalanceGameResults({
         roomId: roomId,
-        round:
-          round.currentRound + 1 === round.totalRounds
-            ? undefined
-            : round.currentRound,
+        round: round.currentRound,
       })
         .then((data) => {
-          //@todo: 데이터 밀리는 현상 백엔드에 확인 후 콘솔 삭제
-          console.log(data); // 이전 데이터
-          console.log(roomId, round); // 최신데이터
           if (data) {
-            if (data.length === 1) {
-              setResult(data);
-            } else {
-              console.log('finalresult 실행');
-              const finalResult: BarProps[] = data.map((data) => ({
-                candidate1: data.a,
-                candidate2: data.b,
-                votes1: data.result.a.length,
-                votes2: data.result.b.length,
-              }));
-              setFinalResult(finalResult);
-            }
+            setResult(data);
           }
         })
         .catch((error) => {
           console.log(error);
         });
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isTimeout]);
 
   const handleSelect = (option: string) => {
@@ -83,11 +64,10 @@ const BalanceGameProgress = ({
     <main className="flex flex-col items-center justify-center p-8 h-full">
       <section className="w-full mb-4 flex flex-col items-center gap-4">
         <Timer
-          //@todo: 테스트 끝나면 원상 복구
           startTime={Date.now()}
           endTime={Date.now() + 3000}
-          //startTime={round.startTime}
-          //endTime={round.endTime}
+          // startTime={round.startTime}
+          // endTime={round.endTime}
           setIsTimeout={setIsTimeout}
         />
       </section>
